Splits shared code out into DiffHelper
