+++ spec/spec_helper.rb
@@ -51,17 +51,3 @@ end
 require 'zonebie/rspec'
 
 RSpec::Expectations.configuration.on_potential_false_positives = :nothing
-
-# Shared helper methods used in multiple files
-def assert_error_messages_equal(err, expected)
-  actual = normalize_error_message(err.message)
-  expected = normalize_error_message(expected)
-  expect(actual).to eql(expected)
-end
-
-def normalize_error_message(message)
-  message.
-    gsub(/\x1b\[[0-9;]*m/, ''). # Strip ANSI control characters used for color
-    gsub(/:0x[0-9a-f]{16}/, ':<id>').
-    strip
-end
